extracts graph component
